eye path is renamed to metrics
