+++ CMakeLists.txt
@@ -271,7 +271,6 @@ add_subdirectory(coding)
 add_subdirectory(drape)
 add_subdirectory(drape_frontend)
 add_subdirectory(editor)
-add_subdirectory(eye)
 add_subdirectory(generator/mwm_diff)
 add_subdirectory(geocoder)
 add_subdirectory(geometry)
@@ -279,6 +278,7 @@ add_subdirectory(indexer)
 add_subdirectory(kml)
 add_subdirectory(local_ads)
 add_subdirectory(map)
+add_subdirectory(metrics)
 add_subdirectory(partners_api)
 add_subdirectory(platform)
 add_subdirectory(routing)
