+++ metrics/metrics_tests_support/eye_for_testing.cpp
@@ -1,7 +1,7 @@
-#include "eye/eye_tests_support/eye_for_testing.hpp"
+#include "metrics/metrics_tests_support/eye_for_testing.hpp"
 
-#include "eye/eye.hpp"
-#include "eye/eye_storage.hpp"
+#include "metrics/eye.hpp"
+#include "metrics/eye_storage.hpp"
 
 #include "coding/internal/file_data.hpp"
