+++ metrics/metrics_tests_support/eye_for_testing.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "eye/eye_info.hpp"
+#include "metrics/eye_info.hpp"
 
 namespace eye
 {
